32D10

+++ src/32D10.h
@@ -3,7 +3,7 @@
 
 #include "global.h"
 
-void func_80032110(u8, s32);
+void func_80032110(u8, s16);
 void func_80032244(u8, s16);
 void func_8003260C(unk_D_86002F58_004_000*);
